+++ src/yafowil/widget/datetime/widget.py
@@ -13,7 +13,7 @@ from yafowil.utils import (
     cssid,
     cssclasses,
     css_managed_props,
-    managedprops,    
+    managedprops,
 )
 from bda.intellidatetime import (
     convert,
@@ -21,16 +21,21 @@ from bda.intellidatetime import (
     DateTimeConversionError,
 )
 
+def call_if_callable(key, widget, data):
+    attr = widget.attrs[key]
+    if callable(attr):
+        return attr(widget, data)
+    return attr
 
 def datetime_extractor(widget, data):
     time = None
-    if widget.attrs['time']:
+    if call_if_callable('time', widget, data):
         time = data.request.get('%s.time' % widget.dottedpath)
     if not widget.attrs['required'] and not data.extracted and not time:
         return ''
+    locale = call_if_callable('locale', widget, data)
+    tzinfo = call_if_callable('tzinfo', widget, data)
     try:
-        locale = widget.attrs['locale']
-        tzinfo = widget.attrs['tzinfo']
         return convert(data.extracted, time=time, tzinfo=tzinfo, locale=locale)
     except DateTimeConversionError:
         raise ExtractionError('Not a valid date input.')
@@ -76,21 +81,19 @@ def render_datetime_input(widget, data, date, time):
         'type': 'text',
         'value':  date,
         'name_': widget.dottedpath,
-        'id': cssid(widget, 'input'),    
-        'class_': cssclasses(widget, data, additional=additional_classes),  
+        'id': cssid(widget, 'input'),
+        'class_': cssclasses(widget, data, additional=additional_classes),
         'size': 10,
-    }        
+    }
     return tag('input', **attrs) + timeinput
 
 
 @managedprops('locale', *css_managed_props)
 def datetime_edit_renderer(widget, data):
-    locale = widget.attrs['locale']
-    delim = widget.attrs['delimiter']
-    if callable(locale):
-        locale = locale(widget, data)
+    locale = call_if_callable('locale', widget, data)
+    delim = call_if_callable('delimiter', widget, data)
+    time = call_if_callable('time', widget, data)
     date = None
-    time = widget.attrs['time']
     if data.value and isinstance(data.value, datetime):
         date = format_date(data.value, locale, delim)
         if time:
@@ -104,25 +107,25 @@ def datetime_edit_renderer(widget, data):
     return render_datetime_input(widget, data, date, time)
 
 
-def render_datetime_display(widget, data, value):
-    if not value:
+def datetime_display_renderer(widget, data):
+    if not data.value:
         return u''
     format = widget.attrs['format']
+    if callable(format):
+        value = format(widget, data)
+    else:
+        value = data.value.strftime(format)
     attrs = {
         'id': cssid(widget, 'display'),
         'class_': 'display-%s' % widget.attrs['class']
     }
-    return data.tag('div', value.strftime(format), **attrs)
-
-
-def datetime_display_renderer(widget, data):
-    return render_datetime_display(widget, data, data.value)
+    return data.tag('div', value, **attrs)
 
 
 factory.register(
-    'datetime', 
+    'datetime',
     extractors=[generic_extractor, generic_required_extractor,
-                datetime_extractor], 
+                datetime_extractor],
     edit_renderers=[datetime_edit_renderer],
     display_renderers=[datetime_display_renderer])
 
@@ -147,27 +150,40 @@ factory.doc['props']['datetime.datepicker'] = \
 factory.defaults['datetime.time'] = False
 factory.doc['props']['datetime.time'] = \
 """Flag whether time input should be rendered.
+
+``time`` may be a callable taking widget and data as parameters expect to return 
+a boolean.
 """
 
 factory.defaults['datetime.tzinfo'] = None
 factory.doc['props']['datetime.tzinfo'] = \
 """Python datetime tzinfo object.
+
+``tzinfo`` may be a callable taking widget and data as parameters expect to 
+return a tzinfo instance.
 """
 
 factory.defaults['datetime.locale'] = 'iso'
 factory.doc['props']['datetime.locale'] = \
-"""Date input format locale. ``yafowil.widget.array`` uses
+"""Date input format locale. ``yafowil.widget.datetime`` uses
 `bda.intellidatetime <http://pypi.python.org/pypi/bda.intellidatetime/>`_ for
 input parsing. Take a look at this package for available locales.
+
+``locale`` may be a callable taking widget and data as parameters expect to 
+return a locale string.
 """
 
-factory.defaults['datetime.delimiter'] = '.'
+factory.defaults['datetime.delimiter'] = '-'
 factory.doc['props']['datetime.delimiter'] = \
 """Delimiter used to render date in input field.
+
+``delimiter`` may be a callable taking widget and data as parameters expect to 
+return a delimiter string.
 """
 
-factory.defaults['datetime.format'] = '%Y.%m.%d - %H:%M'
+factory.defaults['datetime.format'] = '%Y-%m-%d %H:%M'
 factory.doc['props']['datetime.format'] = \
-"""Pattern accepted by ``datetime.strftime``. Used if widget mode is
+"""Pattern accepted by ``datetime.strftime`` or callable taking widget and 
+data as parameters returning unicode or utf-8 string. Used if widget mode is 
 ``display``.
 """